requirements.txt
# MetaTrader5 Pairs Trading Requirements

# Core MT5 integration
MetaTrader5>=5.0.45
ctrader_open_api

# Data analysis and statistics
numpy>=1.21.0
pandas>=1.3.0
statsmodels>=0.13.0
scipy>=1.7.0

# Visualization and reporting
matplotlib>=3.4.0
seaborn>=0.11.0
xlsxwriter>=3.0.0

# Configuration and utilities
python-dotenv>=0.19.0
dataclasses>=0.6

# Optional performance enhancements
numba>=0.56.0
cython>=0.29.0


# Backtesting and simulation
vectorbt